Clases reajustadas
