+++ resources/js/Components/NavLink.jsx
@@ -5,11 +5,9 @@ export default function NavLink({ active = false, className = '', children, ...p
         <Link
             {...props}
             className={
-                'w-44 h-auto font-normal inline-flex justify-center items-center text-xl transition duration-500 ease-in-out ' +
                 (active
-                    ? ' bg-[#efb810] text-black border-b-4 border-black '
-                    : ' text-[#efb810] bg-black hover:bg-[#efb810] hover:text-black') +
-                className
+                    ? ' linkNavActivo'
+                    : ' linkNav')
             }
         >
             {children}
